requirements.txt
    # --- Framework para la API ---
    fastapi
    uvicorn[standard]
    python-multipart

    # --- Base de Datos ---
    sqlalchemy
    psycopg2-binary
    alembic

    # --- Tareas en Segundo Plano ---
    celery
    redis

    # --- Conectores de Modelos LLM ---
    google-generativeai
    openai
    httpx

    # --- Sistema RAG y Vectores ---
    sentence-transformers
    faiss-cpu
    numpy

    # --- Langchain ---
    langchain
    langchain-community

    # --- Utilidades y Seguridad ---
    python-dotenv
    pydantic-settings
    passlib[bcrypt]
    python-jose[cryptography]

    langchain-google-genai
